Added commands to manage game roles

+++ src/DAL/Database/Model/GameRole.cs
@@ -6,7 +6,7 @@ namespace HoU.GuildBot.DAL.Database.Model
     public partial class GameRole
     {
         public short GameRoleID { get; set; }
-        public long DiscordRoleID { get; set; }
+        public decimal DiscordRoleID { get; set; }
         public string RoleName { get; set; }
         public short GameID { get; set; }
         public int ModifiedByUserID { get; set; }
